pyproject.toml
[project]
name = "its-agents-system"
version = "0.1.0"
description = "ITS Multi-Agent AI System for Odoo Operations"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "httpx>=0.25.0",
    "google-generativeai>=0.8.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
    "apscheduler>=3.10.0",
    "python-telegram-bot>=20.0",
    "paramiko>=3.4.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
include = ["agents*", "api*", "common*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py", "*_test.py", "*.spec.py"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
